+++ json_ref_dict/ref_pointer.py
@@ -9,6 +9,8 @@ from json_ref_dict.exceptions import DocumentParseError
 from json_ref_dict.loader import get_document
 from json_ref_dict.uri import URI, parse_segment
 
+UriValuePair = Tuple[URI, Union[List, Dict, None, bool, str, int, float]]
+
 
 class RefPointer(JsonPointer):
     """Subclass of `JsonPointer` which accepts full URI.
@@ -31,31 +33,70 @@ class RefPointer(JsonPointer):
         :return: tuple indicating (1) if doc was a ref and (2) what
             that ref was.
         """
+        remote_uri, value = self.resolve_remote_with_uri(doc, parts_idx)
+        return remote_uri is not None, value
+
+    def resolve_remote_with_uri(
+        self, doc: Any, parts_idx: int
+    ) -> Tuple[Optional[URI], Optional[Any]]:
+        """Defer resolution of references to a new RefPointer.
+
+        :param doc: document element.
+        :param parts_idx: index of the reference part reached.
+        :return: tuple indicating (1) if doc was a ref (the ref URI returned)
+            and (2) what that ref value was.
+        """
         if not (
             isinstance(doc, abc.Mapping) and isinstance(doc.get("$ref"), str)
         ):
-            return False, None
+            return None, None
         remote_uri = self.uri.relative(doc["$ref"]).get(
             *[parse_segment(part) for part in self.parts[parts_idx + 1 :]]
         )
-        return True, resolve_uri(remote_uri)
+
+        resolved_remote_uri, value = resolve_remote_uri(remote_uri)
+        return resolved_remote_uri, value
 
     def resolve(self, doc: Any, default: Any = _nothing) -> Any:
         """Resolves the pointer against doc and returns the referenced object.
 
         If any remotes are found, then resolution is deferred to a new
-        pointer instance in that reference scope.
+        pointer instance in that reference scope and resolving continued,
+        until the value is found.
 
-        :param doc: The document in which to resolve the pointer.
+        :param doc: The document in which to start resolving the pointer.
         :param default: The value to return if the pointer fails. If not
             passed, an exception will be raised.
-        :return: The value of the pointer in this document.
+        :return: The value of the pointer in the containing document.
+            The document may be different from the one given in as an argument.
         :raises JsonPointerException: if `default` is not set and pointer
             could not be resolved.
         """
-        has_remote, remote = self.resolve_remote(doc, -1)
-        if has_remote:
-            return remote
+        _, value = self.resolve_with_uri(doc, default)
+        return value
+
+    get = resolve
+
+    def resolve_with_uri(
+        self, doc: Any, default: Any = _nothing
+    ) -> Tuple[URI, Any]:
+        """Resolves the pointer, starting from given doc
+
+        The resolver recurses references and, as a result, may end up on
+        a different document than the one given in as an argument. The URI for
+        the document containing the value is returned together with the value.
+
+        :param doc: The document with which to start resolving the pointer.
+        :param default: The value to return if the pointer fails. If not
+            passed, an exception will be raised.
+        :return: The updated URI and the value of the pointer
+            in the located document.
+        :raises JsonPointerException: if `default` is not set and pointer
+            could not be resolved.
+        """
+        remote_uri, remote = self.resolve_remote_with_uri(doc, -1)
+        if remote_uri is not None:
+            return remote_uri, remote
         for idx, part in enumerate(self.parts):
             if not part:
                 continue
@@ -64,16 +105,14 @@ class RefPointer(JsonPointer):
                     doc = self.walk(doc, part)
                 except JsonPointerException:
                     doc = self.walk(doc, unquote(part))
-                has_remote, remote = self.resolve_remote(doc, idx)
-                if has_remote:
-                    return remote
+                remote_uri, remote = self.resolve_remote_with_uri(doc, idx)
+                if remote_uri is not None:
+                    return remote_uri, remote
             except JsonPointerException:
                 if default is _nothing:
                     raise
-                return default
-        return doc
-
-    get = resolve
+                return self.uri, default
+        return self.uri, doc
 
     def set(self, doc: Any, value: Any, inplace: bool = True) -> NoReturn:
         """`RefPointer` is read-only."""
@@ -90,10 +129,12 @@ class RefPointer(JsonPointer):
 
 
 @lru_cache(maxsize=None)
-def resolve_uri(uri: URI) -> Union[List, Dict, None, bool, str, int, float]:
+def resolve_remote_uri(
+    uri: URI
+) -> Tuple[URI, Union[List, Dict, None, bool, str, int, float]]:
     """Find the value for a given URI.
 
-    Loads the document and resolves the pointer, bypsasing refs.
+    Loads the document and resolves the pointer, bypassing refs.
     Utilises `lru_cache` to avoid re-loading multiple documents.
     """
     try:
@@ -102,4 +143,12 @@ def resolve_uri(uri: URI) -> Union[List, Dict, None, bool, str, int, float]:
         raise DocumentParseError(
             f"Failed to load base document of {uri}."
         ) from exc
-    return RefPointer(uri).resolve(document)
+
+    pointer = RefPointer(uri)
+    remote_uri, value = pointer.resolve_with_uri(document)
+    return remote_uri if remote_uri else uri, value
+
+
+def resolve_uri(uri: URI) -> Union[List, Dict, None, bool, str, int, float]:
+    _, value = resolve_remote_uri(uri)
+    return value